+++ inst/save_config.php
@@ -1,7 +1,5 @@
 <?php
 
-
-
 define('DIR_BASE', '../');
 require_once(DIR_BASE.'cfg/config.php');
 require_once(DIR_LIB.'generic_lib.php');